Resolve: Remove "comment" key from Color.json
It was added when mapping colors.json to correct videos and images so
that the number of object that have been mapped can be remembered.
However, it remained after the mapping was done and was making a bug in the "Guess color" game.
This was due to the fact that the game considered all keys to be color
names and hence was considering it for the video options. Since the
field do not have any video, this resulted in a blank option.
Hence, it was removed to better user experience.